Add debug information and handle out-of-bounds case in match_vectored
